pyproject.toml
[project]
name = "newen-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "anyio==4.10.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.5",
    "bs4==0.0.2",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "colorama==0.4.6",
    "comm==0.2.3",
    "db-dtypes==1.4.3",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "editorconfig==0.17.1",
    "executing==2.2.1",
    "fake-http-header==0.3.5",
    "fake-useragent==2.2.0",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "gcsfs==2025.7.0",
    "google-api-core==2.25.1",
    "google-api-python-client==2.179.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "google-cloud-bigquery==3.36.0",
    "google-cloud-bigquery-storage==2.33.1",
    "google-cloud-core==2.4.3",
    "google-cloud-storage==3.3.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.1.1",
    "grpcio==1.74.0",
    "grpcio-status==1.74.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "ipykernel==6.30.1",
    "ipython==9.5.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jsbeautifier==1.15.4",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "lingua-language-detector==2.1.1",
    "loguru==0.7.3",
    "matplotlib-inline==0.1.7",
    "multidict==6.6.4",
    "nest-asyncio==1.6.0",
    "numpy==1.26.4",
    "oauthlib==3.3.1",
    "opencv-python==4.5.5.64",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pandas==2.3.2",
    "pandas-gbq==0.28.0",
    "parso==0.8.5",
    "platformdirs==4.4.0",
    "playwright==1.48.0",
    "playwright-stealth==1.0.6",
    "polars-lts-cpu==1.32.3",
    "prompt-toolkit==3.0.52",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydata-google-auth==1.9.1",
    "pyee==12.0.0",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.2",
    "pyzmq==27.1.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "selenium==4.35.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.8",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "tenon==0.1.7",
    "tornado==6.5.2",
    "traitlets==5.14.3",
    "trio==0.30.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.14.1",
    "tzdata==2025.2",
    "undetected-chromedriver==3.5.5",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "wcwidth==0.2.14",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "win32-setctime==1.2.0",
    "wsproto==1.2.0",
    "yarl==1.20.1",
]
